+++ spec/entity/CoverallsReport.spec.php
@@ -11,14 +11,14 @@
 
 namespace coverallskit\spec;
 
-use coverallskit\entity\Report;
+use coverallskit\entity\CoverallsReport;
 use coverallskit\entity\GitRepository;
 use coverallskit\entity\SourceFile;
 use coverallskit\entity\collection\SourceFileCollection;
 use Prophecy\Prophet;
 
 
-describe('Report', function() {
+describe('CoverallsReport', function() {
     beforeEach(function() {
         $this->path = __DIR__ . '/tmp/coverage.json';
 
@@ -31,20 +31,20 @@ describe('Report', function() {
     describe('isEmpty', function() {
         context('when empty', function() {
             it('should return true', function () {
-                $report = new Report();
+                $report = new CoverallsReport();
                 expect($report->isEmpty())->toBeTrue();
             });
         });
     });
     describe('token', function() {
         it('should return repository token string', function() {
-            $report = new Report([ 'token' => 'foo' ]);
+            $report = new CoverallsReport([ 'token' => 'foo' ]);
             expect($report->getToken())->toBe('foo');
         });
     });
     describe('repository', function() {
         it('should return repository', function() {
-            $report = new Report([
+            $report = new CoverallsReport([
                 'repository' => new GitRepository(__DIR__ . '/../../')
             ]);
             expect($report->getRepository())->toBeAnInstanceOf('coverallskit\entity\RepositoryEntity');
@@ -52,7 +52,7 @@ describe('Report', function() {
     });
     describe('sourceFiles', function() {
         it('should return sources file collection', function() {
-            $report = new Report([
+            $report = new CoverallsReport([
                 'sourceFiles' => new SourceFileCollection()
             ]);
             expect($report->getSourceFiles())->toBeAnInstanceOf('coverallskit\entity\collection\SourceFileCollection');
@@ -62,7 +62,7 @@ describe('Report', function() {
     describe('#validate', function() {
         context('when token empty', function() {
             beforeEach(function() {
-                $this->report = new Report([ 'token' => null ]);
+                $this->report = new CoverallsReport([ 'token' => null ]);
             });
             it('throw coverallskit\exception\RequiredException', function() {
                 expect(function() {
@@ -77,7 +77,7 @@ describe('Report', function() {
                 $service = $this->prophet->prophesize('coverallskit\entity\ServiceEntity');
                 $service->isEmpty()->willReturn(true);
 
-                $this->report = new Report([
+                $this->report = new CoverallsReport([
                     'token' => 'foo',
                     'service' => $service->reveal()
                 ]);
@@ -95,7 +95,7 @@ describe('Report', function() {
                 $service = $this->prophet->prophesize('coverallskit\entity\ServiceEntity');
                 $service->isEmpty()->willReturn(false);
 
-                $this->report = new Report([
+                $this->report = new CoverallsReport([
                     'token' => 'foo',
                     'service' => $service->reveal(),
                     'sourceFiles' => new SourceFileCollection()
@@ -125,7 +125,7 @@ describe('Report', function() {
                 'service_name' => 'travis-ci'
             ]);
 
-            $this->report = new Report([
+            $this->report = new CoverallsReport([
                 'token' => 'foo',
                 'repository' => new GitRepository(__DIR__ . '/../../'),
                 'service' => $service->reveal(),
@@ -164,7 +164,7 @@ describe('Report', function() {
                     'service_name' => 'travis-ci'
                 ]);
 
-                $this->notSavedReport = new Report([
+                $this->notSavedReport = new CoverallsReport([
                     'token' => 'foo',
                     'repository' => new GitRepository(__DIR__ . '/../../'),
                     'service' => $service->reveal(),
@@ -181,7 +181,7 @@ describe('Report', function() {
                 unlink($this->notSavedReport->getName());
             });
             it('should use the default name', function() {
-                expect($this->notSavedReport->getName())->toEqual(getcwd() . '/' . Report::DEFAULT_NAME);
+                expect($this->notSavedReport->getName())->toEqual(getcwd() . '/' . CoverallsReport::DEFAULT_NAME);
             });
             it('upload the report file', function() {
                 $this->prophet->checkPredictions();
@@ -203,7 +203,7 @@ describe('Report', function() {
                     'service_name' => 'travis-ci'
                 ]);
 
-                $this->savedReport = new Report([
+                $this->savedReport = new CoverallsReport([
                     'token' => 'foo',
                     'repository' => new GitRepository(__DIR__ . '/../../'),
                     'service' => $service->reveal(),